refactor: 🎨 working on removing redundant and dead code
